requirements.txt
pyyaml
python-box
ensure
joblib
pymupdf
python-docx
pillow            
langchain
langchain-text-splitters
numpy
torch
faiss-cpu
-e .